refactor(build): split the stream readers and the secrets out of build/mod.rs
The build module passed the 500-line limit with the board work, and so did
tests/reporting_contract.rs. The STEP and image-id readers and the failure
replay now live in build/steps.rs, the build.secrets resolution in
build/secrets.rs, and the two build contract tests in tests/build_contract.rs
with the same harness. No behaviour changes.
